Add claim card components
